+++ test/openacc/array1/original_code.f90
@@ -19,7 +19,7 @@ SUBROUTINE claw
     vec1(j) = j
   END DO
 
-  !$claw array-transform parallel acc(loop)
+  !$claw expand parallel acc(loop)
   vec1(:) = vec1(:) + 10
 
   PRINT*,vec1