+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to the CLAW Compiler project are documented in this file.
 * low-level: block directive like `loop-hoist` can now be nested on the same
   depth.
 * driver: `_CRAYFTN` macro is passed directly when Cray preprocessor is used.
+* `array-transform` directive as been renamed `expand` as specified in v2.0 of
+  the CLAW Directive Language Specification.
 * OMNI Compiler submodule now pointing to
   omni-compiler/xcodeml-tools@6dfda5a2bb16ed04487d9e3c688dbba27072944e